update logger to NLog
