--- Main/Loggers/ILogger.cs
@@ -1,10 +1,0 @@
-﻿namespace Arcam.Main.Loggers
-{
-    internal interface ILogger
-    {
-        public void Debug(object data);
-        public void Debug(string symbol, object data);
-        public void Log(object data, string type);
-        public void Log(string symbol, object data, string type);
-    }
-}
