+++ Main/Loggers/TelegramLogger.cs
@@ -1,32 +1,18 @@
-﻿namespace Arcam.Main.Loggers
+﻿using NLog.Targets;
+
+namespace Arcam.Main.Loggers
 {
-    class TelegramLogger : ILogger
+    class TelegramLogger : Target
     {
         public static TelegramBot bot = TelegramBot.getInstance();
-        private readonly string baseName;
-        private readonly string format = "HH:mm:ss.ffffzzz";
-        public static bool isDebug = false;
 
-        public TelegramLogger(string name)
+        protected override void Write(LogEventInfo logEvent)
         {
-            baseName = name;
-        }
-
-        public void Debug(object data)
-        {
-            bot.SendTextMessage($"[{baseName} {DateTime.Now.ToString(format)}]-(Debug) \"{data}\"");
-        }
-        public void Debug(string symbol, object data)
-        {
-            bot.SendTextMessage($"[{baseName} {DateTime.Now.ToString(format)}]-(Debug {symbol}) \"{data}\"");
-        }
-        public void Log(object data, string type)
-        {
-            bot.SendTextMessage($"[{baseName} {DateTime.Now.ToString(format)}]-({type}) \"{data}\"");
-        }
-        public void Log(string symbol, object data, string type)
-        {
-            bot.SendTextMessage($"[{baseName} {DateTime.Now.ToString(format)}]-({type} {symbol}) \"{data}\"");
+            try
+            {
+                bot.SendTextMessage($"{logEvent.LoggerName} {logEvent.FormattedMessage}");
+            }
+            catch { }
         }
     }
 }
